requirements.txt
aiogram==3.7.0
python-dotenv==1.0.1
cryptography==42.0.8
loguru==0.7.2
playwright>=1.44.0
aiofiles>=23.2.1
